FEAT : 보스 체력바 추가, maxGameTime 진입시 게임시간 오르지 않게 변경, maxGameTime 진입시 일반몹 Spawn 되지 않게 변경
