+++ Assets/Undead Survivor/Sprites/Run 0.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 2fe5278f623bb4891a0e277f414b5df4
+guid: 44af9e17ccf8e40da8bd0223395f3e1f
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
